Add useTrashActions hook with dialog management for restoring and deleting parking spots
